+++ src/main/java/nextstep/courses/domain/session/Sessions.java
@@ -1,9 +1,14 @@
 package nextstep.courses.domain.session;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class Sessions {
     private final List<Session> value;
+
+    public Sessions() {
+        this(new ArrayList<>());
+    }
 
     public Sessions(List<Session> value) {
         this.value = value;
